+++ circuits/CHANGELOG.md
@@ -18,6 +18,7 @@ verification keys break backwards compatibility.
 * Variable-length SHA256 gadget [#98](https://github.com/midnightntwrk/midnight-zk/pull/98)
 * Add Schnorr signature example [#55](https://github.com/midnightntwrk/midnight-zk/pull/55)
 * Feature to run internal benchmarks [#93](https://github.com/midnightntwrk/midnight-zk/pull/93)
+* Add function to return number of points involved in a proof [#102](https://github.com/midnightntwrk/midnight-zk/pull/102)
 
 ### Changed
 * Refactored crate to prepare for edition 2024.